Divise le service en 2 comme il est utilisé dans l'ancien frontend
