+++ 2024-frontend/src/components/CanteensTableSatellites.vue
@@ -3,7 +3,7 @@ import { computed } from "vue"
 import { computedAsync } from "@vueuse/core"
 import { useRootStore } from "@/stores/root"
 import { formatSiretOrSiren } from "@/utils"
-import diagnosticService from "@/services/diagnostics.js"
+import diagnosticService from "@/services/diagnosticsBadge.js"
 import campaignService from "@/services/campaigns.js"
 import managersService from "@/services/managers.js"
 import canteensTableService from "@/services/canteensTable.js"
